+++ lib/legion/extensions/apollo.rb
@@ -9,6 +9,9 @@ require 'legion/extensions/apollo/runners/expertise'
 require 'legion/extensions/apollo/runners/maintenance'
 require 'legion/extensions/apollo/runners/entity_extractor'
 require 'legion/extensions/apollo/runners/gas'
+require 'legion/extensions/apollo/runners/request'
+
+require 'legion/extensions/apollo/api' if defined?(Sinatra)
 
 if defined?(Legion::Transport)
   require 'legion/extensions/apollo/transport/exchanges/apollo'